pyproject.toml
[tool.poetry]
name = "south-africa-street-history-mapping"
version = "0.1.0"
description = ""
authors = ["Emily-Rose Steyn <emilyrosesteyn@gmail.com>"]
license = "GPL3.0"
readme = "README.md"
packages = [
    { include = "dictionary_builder", from = "src" },
    { include = "street_list_download", from = "src" },
    { include = "street_list_preprocessing", from = "src" },
    { include = "utils", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^0.14.3"
matplotlib = "^3.8.2"
numpy = "^1.26.4"
plotnine = "^0.12.4"
jupyter = "^1.0.0"
osmnx = "^1.9.1"
osmium = "^3.7.0"
pyrosm = "^0.6.2"
pyarrow = "^15.0.0"
mpi4py = "3.1.3"
python-dotenv = "^1.0.1"
thefuzz = "^0.22.1"
tqdm = "^4.66.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
